Reste des Fonctions avancées +Jpa + Long

+++ ConsomiTounsi/src/main/java/tn/esprit/spring/Service/ADDServiceImpl.java
@@ -32,7 +32,7 @@ public class ADDServiceImpl implements IADDService {
 	}*/
 
 	@Override
-	public void deleteADD(int i) {
+	public void deleteADD(Long i) {
 		AddRepository.deleteById(i);
 		
 	}
