+++ ConsomiTounsi/src/main/java/tn/esprit/spring/Service/CommandServiceImpl.java
@@ -31,7 +31,7 @@ public class CommandServiceImpl implements ICommandService{
 	
 	
 	
-	public void affecterClientACommand(int id, int reference) {
+	public void affecterClientACommand(Long id, int reference) {
 		Command command = commandRepository.findById(reference).get();
 		User user =userRepository.findById(id).get();
 		if (!ObjectUtils.isEmpty(command) && !ObjectUtils.isEmpty(user))
