+++ ConsomiTounsi/src/main/java/tn/esprit/spring/Service/IProductService.java
@@ -7,7 +7,7 @@ import tn.esprit.spring.entities.*;
 public interface IProductService {
 	public 	String addProduct(Product p);
 
-	public void deleteProduct(int i);
+	public void deleteProduct(Long i);
 	
 	public List<Product> retrieveAllProducts();
